* Fix a diff from upstream that was lost in the merge (removal of maxLengthChanged var)
* Fix bug where CM crashes when opening an inline editor near the bottom of
a file / bottom of the viewport (anytime the last rendered line has an inline
open?). There was an incorrect assumption in updateGutter() that indices in
gutterText.childNodes[] correspond one-to-one with line numbers, which isn't
true when inline editor placeholders are present. This wasn't usually a
problem until marijnh/CodeMirror2#548 landed, which gives an NPE when it
grabs an inline placeholder and assumes it has children (although, it looks
like there would also have been subtler bugs if we ever started using gutter
markers for anything). The fix is to track line numbers separately from
child indices (we could instead track the number of inlines and use that
to offset the old "i" counter when setting normalNode, but it means 1 extra
var maintained by the loop either way, and this seems clearer).